お絵かきボードについてIEで表示を拡大縮小したときに描画した内容を保存すると描画内容が歪んでしまうバグを修正 svgからcanvasへ変換するときのparseIntをparseFloatに変更

+++ hifive-ui-library/WebContent/components/artboard/src/DrawingLogic.js
@@ -124,8 +124,8 @@
 					var preY = firstY;
 					// x,yのデータを同時に取り出すので２つずつカウント
 					for (var index = 2, l = pathDataArray.length; index < l; index += 2) {
-						var x = preX + parseInt(pathDataArray[index]);
-						var y = preY + parseInt(pathDataArray[index + 1]);
+						var x = preX + parseFloat(pathDataArray[index]);
+						var y = preY + parseFloat(pathDataArray[index + 1]);
 						ctx.lineTo(x, y);
 						preX = x;
 						preY = y;
